Repurpose existing site goals settings endpoints in Modules/Analytics_4 to be used by the new site-wide setting only.
